chg: :recycle: rename legacy tests with new naming and improve tests quality
